Big shuffle of fields & docs
