+++ tortoise/query_utils.py
@@ -4,8 +4,8 @@ from typing import Any, Dict, List, Optional, Tuple
 from pypika import Table
 from pypika.terms import Criterion
 
-from tortoise import fields
 from tortoise.exceptions import FieldError, OperationalError
+from tortoise.fields.relational import BackwardFKRelation, ManyToManyField
 
 
 def _process_filter_kwarg(model, key, value) -> Tuple[Criterion, Optional[Tuple[Table, Criterion]]]:
@@ -46,7 +46,7 @@ def _get_joins_for_related_field(
     table_pk = related_field.model._meta.db_pk_field
     related_table_pk = related_field.field_type._meta.db_pk_field
 
-    if isinstance(related_field, fields.ManyToManyField):
+    if isinstance(related_field, ManyToManyField):
         related_table = Table(related_field.field_type._meta.table)
         through_table = Table(related_field.through)
         required_joins.append(
@@ -62,7 +62,7 @@ def _get_joins_for_related_field(
                 == getattr(related_table, related_table_pk),
             )
         )
-    elif isinstance(related_field, fields.BackwardFKRelation):
+    elif isinstance(related_field, BackwardFKRelation):
         related_table = Table(related_field.field_type._meta.table)
         required_joins.append(
             (
